resized input dimensions train tested
