+++ official/vision/beta/projects/yt8m/tasks/__init__.py
@@ -1,0 +1,1 @@
+from official.vision.beta.projects.yt8m.configs import yt8m
