+++ official/vision/beta/projects/yt8m/tasks/yt8m_task.py
@@ -21,8 +21,7 @@ from official.core import input_reader
 from official.core import task_factory
 from official.modeling import tf_utils
 from official.vision.beta.projects.yt8m.dataloaders import yt8m_input
-from official.vision.beta.modeling import factory_3d
-from official.vision.beta.projects.yt8m.yt8m_model import YT8MModel
+from official.vision.beta.projects.yt8m.modeling.yt8m_model import YT8MModel
 from official.vision.beta.projects.yt8m.eval_utils import eval_util
 from official.vision.beta.projects.yt8m.configs import yt8m as yt8m_cfg
 
@@ -31,14 +30,11 @@ from official.vision.beta.projects.yt8m.configs import yt8m as yt8m_cfg
 class YT8MTask(base_task.Task):
   """A task for video classification."""
 
-  def build_model(self, num_classes: int=3862, num_frames: int=32):
+  def build_model(self, num_frames: int=32):
     """Builds video classification model."""
-    common_input_shape = [
-        d1 if d1 == d2 else None
-        for d1, d2 in zip(self.task_config.train_data.feature_sizes,
-                          self.task_config.validation_data.feature_sizes)
-    ]
-    input_specs = tf.keras.layers.InputSpec(shape=[None] + common_input_shape)
+    data_cfg = self.task_config.train_data
+    common_input_shape = [data_cfg.max_frames, sum(data_cfg.feature_sizes)] #TODO: revise
+    input_specs = tf.keras.layers.InputSpec(shape=common_input_shape)
     logging.info('Build model input %r', common_input_shape)
 
     #model configuration
@@ -47,7 +43,7 @@ class YT8MTask(base_task.Task):
               input_params=model_config,
               input_specs=input_specs,
               num_frames=num_frames,
-              num_classes=num_classes
+              num_classes=data_cfg.num_classes
               )
     return model
 
@@ -75,41 +71,48 @@ class YT8MTask(base_task.Task):
       labels: labels.
       model_outputs: Output logits of the classifier.
       aux_losses: auxiliarly loss tensors, i.e. `losses` in keras.Model.
+
     Returns:
-      The total loss tensor.
+      The total loss, model loss tensors.
     """
     losses_config = self.task_config.losses
-    total_loss = tf.keras.losses.binary_crossentropy(
+    model_loss = tf.keras.losses.binary_crossentropy(
       labels,
       model_outputs,
       from_logits=losses_config.from_logits,
       label_smoothing=losses_config.label_smoothing)
 
-    total_loss = tf_utils.safe_mean(total_loss)
-
+    model_loss = tf_utils.safe_mean(model_loss) #TODO: remove?
+    total_loss = model_loss
     if aux_losses:
       total_loss += tf.add_n(aux_losses)
 
-    return total_loss
+    return total_loss, model_loss
 
 
-  def build_metrics(self, num_classes: int=3862, top_k=20, top_n=None, training=True):
+  def build_metrics(self, num_classes: int=3862, training=True):
     """Gets streaming metrics for training/validation.
        metric: mAP/gAP
       Args:
       num_class: A positive integer specifying the number of classes.
+
       top_k: A positive integer specifying how many predictions are considered
         per video.
       top_n: A positive Integer specifying the average precision at n, or None
         to use all provided data points.
     """
-    metrics = eval_util.EvaluationMetrics(num_classes, top_k=top_k, top_n=top_n)
-    return [metrics.map_calculator, metrics.global_ap_calculator]
+    metrics = []
+    metric_names = ['total_loss', 'model_loss']
+    for name in metric_names:
+      metrics.append(tf.keras.metrics.Mean(name, dtype=tf.float32))
 
+    if not training: #cannot run in train step
+      top_k = self.task_config.top_k
+      top_n = self.task_config.top_n
+      self.avg_prec_metric = eval_util.EvaluationMetrics(
+        num_classes, top_k=top_k, top_n=top_n)
 
-  def process_metrics(self, metrics, labels, outputs, loss):
-    '''Processes metrics'''
-    metrics.accumulate(outputs=outputs, labels=labels, loss=loss)
+    return metrics
 
 
   def train_step(self, inputs, model, optimizer, metrics=None):
@@ -125,11 +128,16 @@ class YT8MTask(base_task.Task):
       model: the model, forward pass definition.
       optimizer: the optimizer for this training step.
       metrics: a nested structure of metrics objects.
+
     Returns:
       A dictionary of logs.
     """
     features, labels = inputs['video_matrix'], inputs['labels']
-    # video_ids, num_frames = inputs['video_ids'], inputs['num_frames']
+    features = tf.squeeze(features) #(batch, 1, classes) -> (batch, classes)
+    labels = tf.squeeze(labels)
+    print("---------------- YT8M_TASK.PY ----------------")
+    print("features: ", features.shape)
+    print("---------------- YT8M_TASK.PY ----------------")
 
     num_replicas = tf.distribute.get_strategy().num_replicas_in_sync
     with tf.GradientTape() as tape:
@@ -138,18 +146,21 @@ class YT8MTask(base_task.Task):
       # mixed_float16 or mixed_bfloat16 to ensure output is casted as float32.
       outputs = tf.nest.map_structure(lambda x: tf.cast(x, tf.float32), outputs)
 
+
       # Computes per-replica loss
-      loss = self.build_losses(
+      loss, model_loss = self.build_losses(
         model_outputs=outputs, labels=labels, aux_losses=model.losses)
       # Scales loss as the default gradients allreduce performs sum inside the
       # optimizer.
       scaled_loss = loss / num_replicas
+
 
       # For mixed_precision policy, when LossScaleOptimizer is used, loss is
       # scaled for numerical stability.
       if isinstance(
               optimizer, tf.keras.mixed_precision.experimental.LossScaleOptimizer):
         scaled_loss = optimizer.get_scaled_loss(scaled_loss)
+
 
     tvars = model.trainable_variables
     grads = tape.gradient(scaled_loss, tvars)
@@ -166,35 +177,74 @@ class YT8MTask(base_task.Task):
     optimizer.apply_gradients(list(zip(grads, tvars)))
 
     logs = {self.loss: loss}
+
+    all_losses = {
+      'total_loss': loss,
+      'model_loss': model_loss
+    }
+
     if metrics:
-      info_dict = self.process_metrics(metrics, labels, outputs, loss)
-      logs.update(metrics.get())
-    #TODO: model.compiled_metrics - removed
+      for m in metrics:
+        m.update_state(all_losses[m.name])
+        logs.update({m.name: m.result()})
+
     return logs
 
   def validation_step(self, inputs, model, metrics=None):
     """Validatation step.
+
     Args:
       inputs: a dictionary of input tensors.
+              output_dict = {
+            "video_ids": batch_video_ids,
+            "video_matrix": batch_video_matrix,
+            "labels": batch_labels,
+            "num_frames": batch_frames,
+            }
       model: the keras.Model.
       metrics: a nested structure of metrics objects.
+
     Returns:
       A dictionary of logs.
     """
-    features, labels = inputs
+    features, labels = inputs['video_matrix'], inputs['labels']
 
     outputs = self.inference_step(features['image'], model)
     outputs = tf.nest.map_structure(lambda x: tf.cast(x, tf.float32), outputs) #TODO: check if necessary
-    loss = self.build_losses(model_outputs=outputs, labels=labels,
+    loss, model_loss = self.build_losses(model_outputs=outputs, labels=labels,
                              aux_losses=model.losses)
 
     logs = {self.loss: loss}
+
+    all_losses = {
+      'total_loss' : loss,
+      'model_loss' : model_loss
+    }
+
+    self.avg_prec_metric.accumulate(predictions=outputs, labels=labels)
+    logs.update(metrics.get())
+
     if metrics:
-      info_dict = self.process_metrics(metrics, labels, outputs, loss)
-      logs.update(metrics.get())
-    #TODO: model.compiled_metrics - removed
+      for m in metrics:
+        m.update_state(all_losses[m.name])
+        logs.update({m.name: m.result()})
     return logs
 
   def inference_step(self, inputs, model):
     """Performs the forward step."""
     return model(inputs, training=False)
+
+  def aggregate_logs(self, state=None, step_outputs=None):
+    if state is None:
+      # self.coco_metric.reset_states()
+      self.avg_prec_metric.clear()
+      # state = self.coco_metric
+      state = self.avg_prec_metric
+    # self.coco_metric.update_state(step_outputs[self.coco_metric.name][0],
+    #                               step_outputs[self.coco_metric.name][1])
+    self.avg_prec_metric.accumulate(predictions=step_outputs[self.avg_prec_metric.name][1], labels=step_outputs[self.avg_prec_metric][0])
+    return state
+
+  def reduce_aggregated_logs(self, aggregated_logs):
+    # return self.coco_metric.result()
+    return self.avg_prec_metric.get()
